Add Rack::ResponseCache
The included RDoc gives a good description of the use of this.
Basically, successful responses for GET requests without query
strings are cached either to disk or to a user provided ruby object
(such as a hash or memcache).

In the most basic form, this does basically the same thing as Rails'
page caching, but there is a lot more flexibility, since you can
provide your own cache object as well as the block that maps request
paths to file system paths or cache keys.