Middleware that sets an X-Runtime header
Times how long it takes for the application to respond to the request
and sets that to the X-Runtime header of the response.

Also allows the user to provide a suffix, so that different things can
be timed, eg, the just the app, or the app plug middleware stack.

    use Rack::Runtime, "All"

    # use more middleware's here

    use Rack::Runtime, "App"

    run Application

will set "X-Runtime-All" and "X-Runtime-App" headers on the response.

Signed-off-by: Joshua Peek <josh@joshpeek.com>