Ruby 2.5 support
This...  wasn't fun.  Turns out that older versions of ruby-prof aren't Ruby
2.5-compatible, but at some point between ruby-prof 0.13 and ruby-prof 0.17,
they broke the AbstractPrinter API, which breaks our ability to use those
printers in Rack::Profiler.  I wouldn't have noticed, except one of the
printers that was broken was the one we were using in a test case.

Also bumped all the other Rubies in .travis.yml to the latest version,
because I was in there already.