Rack::JSONBodyParser: rescue all parser exceptions
Rescue all exceptions raised by the parser as long as it is StandardError
subclass (which should be all the exceptions raised by different libraries).

This is specially needed when using a parser other than JSON. Otherwise we need
to rescue the corresponding exception and raise JSON:ParseError inside the
block.