Drop support for Ruby 2.2
git-version-bump can't be used on Ruby 2.2 due to the use of &.
https://github.com/mpalmer/git-version-bump/commit/640fedcd8f9934bef36cf5675bb9b7366c18e099#diff-88985aff4dfd862090e947417b1b4e60694e6028380ac827cae988331a3bb456R291
and the fact that the required ruby version hasn't been updated for the
gem. https://github.com/mpalmer/git-version-bump/blob/c1af65cd82c131cb541fa717b3d24a9247973049/git-version-bump.gemspec#L12
Instead of fighting that battle, let's drop 2.2, it is very old (soon 10
years).

Also, there is no need to rescue the LoadError when GVB is used in the
gemspec, we must be able to load it.